优化&修复 ·	活动时间如果二次编辑，点击X会清空，应该保留上一次的时间·	活动描述那个框，添加暗纹文字“可输入活动描述，让其他人更好了解您的活动内容”·	在编辑活动描述的时候，如果想尝试滑动页面，会被输入法完全遮住编辑区域·	活动描述编辑框经常点击无反应

+++ Metic/Source/FlatDatePicker/FlatDatePicker.h
@@ -90,16 +90,16 @@ typedef NS_ENUM(NSInteger, FlatDatePickerMode) {
     NSMutableArray *_labelsSeconds;
     
     // Date selected :
-    int _selectedDay;
-    int _selectedMonth;
-    int _selectedYear;
+    NSInteger _selectedDay;
+    NSInteger _selectedMonth;
+    NSInteger _selectedYear;
     
-    int _selectedDate;
+    NSInteger _selectedDate;
     
     // Time selected :
-    int _selectedHour;
-    int _selectedMinute;
-    int _selectedSecond;
+    NSInteger _selectedHour;
+    NSInteger _selectedMinute;
+    NSInteger _selectedSecond;
     
     // First init flag :
     BOOL _isInitialized;
